+++ packages/react-core/src/evm/providers/thirdweb-sdk-provider.tsx
@@ -1,20 +1,13 @@
-import { DEFAULT_API_KEY } from "../../core/constants/rpc";
+import { useMemo, useEffect } from "react";
 import { QueryClientProviderWithDefault } from "../../core/providers/query-client";
 import { ThirdwebConfigProvider } from "../contexts/thirdweb-config";
+import { ThirdwebSDKContext } from "../contexts/thirdweb-sdk";
 import { ThirdwebConnectedWalletProvider } from "../contexts/thirdweb-wallet";
-import { useUpdateChainsWithApiKeys } from "../hooks/chain-hooks";
+import { useUpdateChainsWithClientId } from "../hooks/chain-hooks";
 import { ThirdwebSDKProviderProps } from "./types";
-import { Chain, defaultChains, getChainRPC } from "@thirdweb-dev/chains";
+import { Chain, defaultChains, getValidChainRPCs } from "@thirdweb-dev/chains";
 import { SDKOptionsOutput, ThirdwebSDK } from "@thirdweb-dev/sdk/evm";
-import { createContext, useContext, useEffect, useMemo } from "react";
-import invariant from "tiny-invariant";
-
-interface TWSDKContext {
-  sdk?: ThirdwebSDK;
-  _inProvider?: true;
-}
-
-const ThirdwebSDKContext = /* @__PURE__ */ createContext<TWSDKContext>({});
+import { checkClientIdOrSecretKey } from "@thirdweb-dev/sdk";
 
 /**
  *
@@ -27,9 +20,7 @@ const WrappedThirdwebSDKProvider = <TChains extends Chain[]>({
   activeChain,
   signer,
   children,
-  thirdwebApiKey,
-  infuraApiKey,
-  alchemyApiKey,
+  clientId,
 }: React.PropsWithChildren<
   { supportedChains: Readonly<TChains> } & Omit<
     ThirdwebSDKProviderProps<TChains>,
@@ -70,11 +61,7 @@ const WrappedThirdwebSDKProvider = <TChains extends Chain[]>({
 
     if (supportedChain && supportedChain.rpc.length > 0) {
       try {
-        const rpcUrl = getChainRPC(supportedChain, {
-          thirdwebApiKey,
-          infuraApiKey,
-          alchemyApiKey,
-        });
+        const rpcUrl = getValidChainRPCs(supportedChain, clientId)[0];
 
         readonlySettings = {
           chainId: supportedChain.chainId,
@@ -105,7 +92,7 @@ const WrappedThirdwebSDKProvider = <TChains extends Chain[]>({
       // sdk from chainId
       sdk_ = new ThirdwebSDK(
         chainId,
-        { ...mergedOptions, infuraApiKey, alchemyApiKey, thirdwebApiKey },
+        { ...mergedOptions, clientId },
         storageInterface,
       );
     }
@@ -126,15 +113,7 @@ const WrappedThirdwebSDKProvider = <TChains extends Chain[]>({
     (sdk_ as any)._chainId = chainId;
 
     return sdk_;
-  }, [
-    activeChainId,
-    alchemyApiKey,
-    infuraApiKey,
-    supportedChains,
-    sdkOptions,
-    storageInterface,
-    thirdwebApiKey,
-  ]);
+  }, [activeChainId, supportedChains, sdkOptions, storageInterface, clientId]);
 
   useEffect(() => {
     // if we have an sdk and a signer update the signer
@@ -179,21 +158,24 @@ export const ThirdwebSDKProvider = <TChains extends Chain[]>({
   queryClient,
   supportedChains,
   activeChain,
-  thirdwebApiKey = DEFAULT_API_KEY,
-  alchemyApiKey,
-  infuraApiKey,
+  clientId,
   ...restProps
 }: React.PropsWithChildren<ThirdwebSDKProviderProps<TChains>>) => {
+  if (!clientId) {
+    checkClientIdOrSecretKey(
+      "No clientId provided in ThirdwebSDK. You will have limited access to thirdweb's services for storage, RPC, and account abstraction. You can get a clientId from https://thirdweb.com/create-api-key",
+      clientId,
+      undefined,
+    );
+  }
   const supportedChainsNonNull = useMemo(() => {
     return supportedChains || (defaultChains as any as TChains);
   }, [supportedChains]);
   const [supportedChainsWithKey, activeChainIdOrObjWithKey] =
-    useUpdateChainsWithApiKeys(
+    useUpdateChainsWithClientId(
       supportedChainsNonNull,
       activeChain || supportedChainsNonNull[0],
-      thirdwebApiKey,
-      alchemyApiKey,
-      infuraApiKey,
+      clientId,
     );
 
   const mergedChains = useMemo(() => {
@@ -222,61 +204,22 @@ export const ThirdwebSDKProvider = <TChains extends Chain[]>({
     <ThirdwebConfigProvider
       value={{
         chains: mergedChains as Chain[],
-        thirdwebApiKey,
-        alchemyApiKey,
-        infuraApiKey,
+        clientId,
       }}
     >
-      <ThirdwebConnectedWalletProvider signer={signer}>
-        <QueryClientProviderWithDefault queryClient={queryClient}>
-          <WrappedThirdwebSDKProvider
-            signer={signer}
-            supportedChains={mergedChains}
-            thirdwebApiKey={thirdwebApiKey}
-            alchemyApiKey={alchemyApiKey}
-            infuraApiKey={infuraApiKey}
-            activeChain={activeChainIdOrObjWithKey}
-            {...restProps}
-          >
+      <QueryClientProviderWithDefault queryClient={queryClient}>
+        <WrappedThirdwebSDKProvider
+          signer={signer}
+          supportedChains={mergedChains}
+          clientId={clientId}
+          activeChain={activeChainIdOrObjWithKey}
+          {...restProps}
+        >
+          <ThirdwebConnectedWalletProvider signer={signer}>
             {children}
-          </WrappedThirdwebSDKProvider>
-        </QueryClientProviderWithDefault>
-      </ThirdwebConnectedWalletProvider>
+          </ThirdwebConnectedWalletProvider>
+        </WrappedThirdwebSDKProvider>
+      </QueryClientProviderWithDefault>
     </ThirdwebConfigProvider>
   );
 };
-
-/**
- * @internal
- */
-function useSDKContext(): TWSDKContext {
-  const ctx = useContext(ThirdwebSDKContext);
-  invariant(
-    ctx._inProvider,
-    "useSDK must be called from within a ThirdwebProvider, did you forget to wrap your app in a <ThirdwebProvider />?",
-  );
-  return ctx;
-}
-
-/**
- *
- * @returns {@link ThirdwebSDK}
- * Access the instance of the thirdweb SDK created by the ThirdwebProvider
- * to call methods using the connected wallet on the desiredChainId.
- * @example
- * ```javascript
- * const sdk = useSDK();
- * ```
- */
-export function useSDK(): ThirdwebSDK | undefined {
-  const { sdk } = useSDKContext();
-  return sdk;
-}
-
-/**
- * @internal
- */
-export function useSDKChainId(): number | undefined {
-  const sdk = useSDK();
-  return (sdk as any)?._chainId;
-}