+++ packages/sdk/src/evm/contracts/prebuilt-implementations/marketplacev3.ts
@@ -214,6 +214,7 @@ export class MarketplaceV3 implements UpdateableNetwork {
       address,
       abi,
       options,
+      storage,
     ),
   ) {
     this._chainId = chainId;
@@ -254,7 +255,8 @@ export class MarketplaceV3 implements UpdateableNetwork {
    * @internal
    */
   public async prepare<
-    TMethod extends keyof MarketplaceV3Contract["functions"] = keyof MarketplaceV3Contract["functions"],
+    TMethod extends
+      keyof MarketplaceV3Contract["functions"] = keyof MarketplaceV3Contract["functions"],
   >(
     method: string & TMethod,
     args: any[] & Parameters<MarketplaceV3Contract["functions"][TMethod]>,
@@ -272,7 +274,8 @@ export class MarketplaceV3 implements UpdateableNetwork {
    * @internal
    */
   public async call<
-    TMethod extends keyof MarketplaceV3Contract["functions"] = keyof MarketplaceV3Contract["functions"],
+    TMethod extends
+      keyof MarketplaceV3Contract["functions"] = keyof MarketplaceV3Contract["functions"],
   >(
     functionName: string & TMethod,
     args?: Parameters<MarketplaceV3Contract["functions"][TMethod]>,