+++ packages/wallets/src/evm/connectors/magic/index.ts
@@ -21,18 +21,17 @@ import { Magic } from "magic-sdk";
 import type { AbstractProvider } from "web3-core";
 import { RPCProviderModule } from "@magic-sdk/provider/dist/types/modules/rpc-provider";
 
-export type MagicAuthConnectOptions =
+export type MagicAuthConnectOptions = {
+  chainId?: number;
+} & (
   | {
-      chainId?: number;
-    } & (
-      | {
-          email: string;
-        }
-      | {
-          phoneNumber: string;
-        }
-      | {}
-    );
+      email: string;
+    }
+  | {
+      phoneNumber: string;
+    }
+  | {}
+);
 // | {
 //     oauthProvider: OAuthProvider;
 //   }