+++ packages/wallets/src/evm/wallets/base.ts
@@ -2,7 +2,7 @@ import { AsyncStorage, createAsyncLocalStorage } from "../../core/AsyncStorage";
 import type { DAppMetaData } from "../../core/types/dAppMeta";
 import { ConnectParams, Connector } from "../interfaces/connector";
 import { AbstractWallet } from "./abstract";
-import { Chain, defaultChains } from "@thirdweb-dev/chains";
+import { Chain, defaultChains, updateChainRPCs } from "@thirdweb-dev/chains";
 import { DEFAULT_DAPP_META } from "../constants/dappMeta";
 import { EVMWallet } from "../interfaces";
 
@@ -11,6 +11,7 @@ export type WalletOptions<TOpts extends Record<string, any> = {}> = {
   walletId?: string;
   walletStorage?: AsyncStorage;
   dappMetadata?: DAppMetaData;
+  clientId?: string;
 } & TOpts;
 
 export type WalletMeta = {
@@ -43,7 +44,9 @@ export abstract class AbstractClientWallet<
     super();
     this.walletId = walletId;
     this.options = options;
-    this.chains = options?.chains || defaultChains;
+    this.chains = (options?.chains || defaultChains).map((c) =>
+      updateChainRPCs(c, options?.clientId),
+    );
     this.dappMetadata = options?.dappMetadata || DEFAULT_DAPP_META;
     this.walletStorage =
       options?.walletStorage || createAsyncLocalStorage(this.walletId);
@@ -174,9 +177,11 @@ export abstract class AbstractClientWallet<
   }
 
   async updateChains(chains: Chain[]) {
-    this.chains = chains;
+    this.chains = chains.map((c) => {
+      return updateChainRPCs(c, this.options?.clientId);
+    });
     const connector = await this.getConnector();
-    connector.updateChains(chains);
+    connector.updateChains(this.chains);
   }
 
   /**